requirements.txt
beautifulsoup4
requests
faust-cchardet
lxml